creating posts works

+++ models/forum.go
@@ -5,6 +5,6 @@ type Forum struct {
 	Title   string `json:"title"`
 	User    string `json:"user"`
 	Slug    string `json:"slug"`
-	Posts   int    `json:"posts,omitempty"`
+	Posts   int64  `json:"posts,omitempty"`
 	Threads int    `json:"threads,omitempty"`
 }
