+++ models/post.go
@@ -6,15 +6,15 @@ import (
 
 //easyjson:json
 type Post struct {
-	Id       int       `json:"id,omitempty"`
-	Parent   int       `json:"parent,omitempty"`
+	Id       int64     `json:"id,omitempty"`
+	Parent   int64     `json:"parent,omitempty"`
 	Author   string    `json:"author"`
 	Message  string    `json:"message"`
 	IsEdited bool      `json:"isEdited,omitempty"`
 	Forum    string    `json:"forum,omitempty"`
 	Thread   int       `json:"thread,omitempty"`
 	Created  time.Time `json:"created,omitempty"`
-	Path     []int     `json:"-"`
+	Path     []int64   `json:"-"`
 }
 
 //easyjson:json
